add jobs list, fix company  settings and fix the design

+++ app/Http/Controllers/JobController.php
@@ -23,13 +23,7 @@ class JobController extends Controller
                 $package = JobService::checkPackage(Auth::user()->id);
                 $subscription = JobService::checkSubscription(Auth::user()->id);
                 if ($package || $subscription) {
-                    if ($package->number > 0) {
-                        return view('account.job.add');
-                    }
-                    if ($subscription->number > 0 && $subscription->expired_at > now())  {
-                        return view('account.job.add');
-                    }
-                    
+                    return view('account.job.add');
                 }
                 return redirect('/offers');
             }
@@ -48,34 +42,36 @@ class JobController extends Controller
      */
     public function create(Request $request)
     {
-        dd($request);
+        //dd($request->user()->can('create'));
+        $request->user()->can('create');
         $request->validate([
             "image" => "required",
-            "title" => "required",
+            "title" => "required | min:5",
             "type" => "required",
             "location" => "required",
             "description" => "required | min:255",
             "apply_by" => "required",
-            "email_to_apply" => "required",
+            "to_apply" => "required",
         ]);
         try {
-            $path = AppService::saveFile($request->name, $request->file("image"), "jobs");
-            job::create([
+            $path = AppService::saveFile($request->title, $request->file("image"), "jobs");
+            $job = job::create([
                 "title" => $request->title,
                 "description" => $request->description,
-                "status" => $request->status,
+                "type" => $request->type,
                 "location" => $request->location,
                 "image" => $path,
                 "apply_by" => $request->apply_by,
                 "to_apply" => $request->to_apply,
+                "experience" => $request->experience,
+                "expired_at" => $request->expired_at,
+                "company_id" => Auth::user()->id,
             ]);
-            return view('account.job.details');
-            
+            return redirect("/account");
         } catch (\Throwable $th) {
-            //throw $th;
+            dd($th->getMessage());
+            return back()->with('message', 'error, retry and refresh ');
         }
-        
-       
     }
 
 
@@ -88,16 +84,18 @@ class JobController extends Controller
      */
     public function delete($id, Request $request)
     {
-       try {
-        $job = Job::find($id);
-        if ($request->user()->cannot('delete', $job)) {
-            abort('401');
+        try {
+            $job = Job::find($id);
+            if (!$job)
+                abort('404');
+            if ($request->user()->cannot('delete', $job)) {
+                abort('401');
+            }
+            $job->delete();
+            return redirect('/account')->with('message', 'Job successfull delete');
+        } catch (\Throwable $th) {
+            return back()->with('message', 'Error, refresh and try again');
         }
-        $job->delete();
-        return redirect('/account')->with('message', 'Job successfull delete');
-       } catch (\Throwable $th) {
-        return back()->with('message', 'Error, refresh and try again');
-       }
     }
 
 
@@ -110,6 +108,8 @@ class JobController extends Controller
     public function edit($id, Request $request)
     {
         $job = Job::find($id);
+        if (!$job)
+            abort('404');
         if ($request->user()->cannot('edit', $job)) {
             abort('401');
         }
@@ -121,10 +121,12 @@ class JobController extends Controller
     {
         try {
             $job = Job::find($id);
+            if (!$job)
+                abort('404');
             if ($request->user()->cannot('edit', $job)) {
                 abort('401');
             }
-            
+
             $job->title = $request->title;
             $job->type = $request->type;
             $job->location = $request->location;
@@ -137,9 +139,39 @@ class JobController extends Controller
             }
             $job->save();
             return redirect('/account')->with('message', 'Job successfull update');
-
         } catch (\Throwable $th) {
             return back()->with('message', 'Error, refresh and try again');
         }
+    }
+
+
+    public function details($id)
+    {
+        $job =  Job::find($id);
+        if ($job) {
+            $job->view += 1;
+            $job->save();
+            return  view('account.job.details', compact('job'));
+        } else  abort('404');
+    }
+
+
+    public function search(Request $request)
+    {
+        $KeyWord = $request->keyword ?? '';
+        $location = $request->location ?? '';
+        $type = $request->type ?? '';
+        $result  = new Job();
+        if (isset($KeyWord) && !empty($KeyWord)) {
+            $result = $result->where('title', 'like', "%$KeyWord%")->orWhere('description', 'like', "%$KeyWord%");
+        }
+        if (isset($location) && !empty($location)) {
+            $result = $result->where('location', 'like', "%$location%");
+        }
+        if (isset($type) && !empty($type)) {
+            $result = $result->where('type', $type);
+        }
+         $res  = $result->paginate(10);
+         return view("search", ['results' => $res]);
     }
 }
